update lecture #4: rename files
