+++ lectures/04_getting_started_with_nodejs/sms/sms.js
@@ -2,53 +2,48 @@ var http = require("http");
 var querystring = require("querystring");
 
 
-function sendSms(sender, receiver, content) {
-  var requestData = querystring.stringify({
-    "send_phone": sender,
-    "dest_phone": receiver,
-    "msg_body": content
-  });
+var requestData = querystring.stringify({
+  "send_phone": "01022205736",
+  "dest_phone": "01022205736",
+  "msg_body": "hello world"
+});
 
 
-  var requestOptions = {
-    hostname: "api.openapi.io",
-    path: "/ppurio/1/message/sms/dobestan/",
-    method: "POST",
-    headers: {
-      "Content-Type": "application/x-www-form-urlencoded",
-      "Content-Length": requestData.length,
+var requestOptions = {
+  hostname: "api.openapi.io",
+  path: "/ppurio/1/message/sms/dobestan/",
+  method: "POST",
+  headers: {
+    "Content-Type": "application/x-www-form-urlencoded",
+    "Content-Length": requestData.length,
 
-      "x-waple-authorization": "MTkyMC0xNDEzODU0NTAwMzU3LTllM2VkOTM3LTYwMTEtNGU2Zi1iZWQ5LTM3NjAxMTNlNmYyMg==",
-    }
-  };
-
-
-  var request = http.request(
-    requestOptions,
-    function(response) {
-      var data = "";
-
-      response.on("data", function(chunk) {
-        data += chunk;
-      });
-
-      response.on("end", function() {
-        console.log(data);
-      });
-    }
-  );
+    "x-waple-authorization": "MTkyMC0xNDEzODU0NTAwMzU3LTllM2VkOTM3LTYwMTEtNGU2Zi1iZWQ5LTM3NjAxMTNlNmYyMg==",
+  }
+};
 
 
-  request.on("error", function(error) {
-    console.log(error);
-  });
+var request = http.request(
+  requestOptions,
+  function(response) {
+    var data = "";
+
+    response.on("data", function(chunk) {
+      data += chunk;
+    });
+
+    response.on("end", function() {
+      console.log(data);
+    });
+  }
+);
 
 
-  request.write(requestData);
+request.on("error", function(error) {
+  console.log(error);
+});
 
 
-  request.end();
-}
+request.write(requestData);
 
 
-module.exports = sendSms;
+request.end();
